requirements.txt
# =======================
# Numeric / Scientific (base)
# =======================
numpy==1.26.4
scipy==1.11.4

# =======================
# Core Computer Vision
# =======================
# OpenCV 4.10 still works with NumPy 1.26 and Python 3.10
opencv-python==4.10.0.84

# Face detection / embeddings
facenet-pytorch==2.6.0

# Mediapipe with Python 3.10 + NumPy 1.x
mediapipe==0.10.14

# =======================
# Object Detection (YOLO)
# =======================
ultralytics==8.1.0

# =======================
# Machine Learning Framework (PyTorch trio)
# =======================
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# =======================
# Audio Processing
# =======================
# May require: brew install portaudio
pyaudio==0.2.13


# Whisper for STT (this meta-package pins a working combo)
openai-whisper==20231106

# Librosa family (built against NumPy 1.x)
librosa==0.10.1
soundfile==0.12.1

# =======================
# Screen Capture
# =======================
mss==9.0.1

# =======================
# Alert System (TTS / audio playback)
# =======================
gTTS==2.5.1
pygame==2.5.2
playsound==1.2.2

# =======================
# Reporting & Visualisation
# =======================
pdfkit==1.0.0
jinja2==3.1.3
matplotlib==3.7.3
Pillow==10.2.0

# =======================
# System Utilities & Web
# =======================
psutil==5.9.8
pyyaml==6.0.1
flask==3.0.2

# =======================
# Transformers stack (for WavLM speaker embeddings)
# =======================
tokenizers==0.13.3
transformers==4.30.2
huggingface-hub==0.20.3